Extract order-analysis heuristics
