+++ apps/server/README.md
@@ -24,7 +24,7 @@ Backend ownership boundaries:
 - `infra/runtime/`: lifecycle management, processing loop, runtime health state, and WebSocket broadcast coordination.
 - `infra/processing/`: signal processing pipeline.
 - `infra/config/`: runtime settings stores used by recording and runtime services.
-- `use_cases/diagnostics/`: post-stop analysis/findings logic; the package-level API still re-exports shared vehicle-order helpers used by live telemetry.
+- `use_cases/diagnostics/`: post-stop analysis/findings logic; `order_analysis.py` now keeps the core matching/session engine while `order_heuristics.py` owns the pure scoring/filter heuristics, and the package-level API still re-exports shared vehicle-order helpers used by live telemetry.
 - `use_cases/run/`: recording orchestration; `lifecycle_state.py` owns the in-memory run session state, `persistence_writer.py` owns history-write coordination and retry bookkeeping, `logger.py` owns the coordinator, and `post_analysis.py` owns the background analysis queue above the injected persistence/error boundary.
 - `adapters/persistence/`, `shared/`, and `use_cases/history/`: SQLite persistence, shared typed contracts/pure helpers (including run-log decoding, the `RunPersistence`, `ClientTracker`, `SignalSource`, `SpeedProvider`, and `SettingsReader` ports, and vehicle-order math), car library loading, and history/report/export services. `use_cases/history/report_interpretation.py` owns pure report-domain interpretation used by PDF mapping, while `adapters/persistence/history_db/` keeps `HistoryDB` as the public facade above internal `_run_lifecycle.py`, `_sample_io.py`, and `_queries.py` sections plus shared `_schema.py` and `_samples.py` helpers.
 - `adapters/pdf/`: PDF/report rendering pipeline.
