+++ apps/server/tests/use_cases/diagnostics/test_confidence_scoring_regressions.py
@@ -14,7 +14,7 @@ import inspect
 import pytest
 from test_support.findings import make_finding
 
-from vibesensor.use_cases.diagnostics.order_analysis import (
+from vibesensor.use_cases.diagnostics.order_heuristics import (
     suppress_engine_aliases as _suppress_engine_aliases,
 )
 from vibesensor.use_cases.diagnostics.order_statistics import (
