+++ apps/server/vibesensor/use_cases/diagnostics/order_analysis.py
@@ -1,4 +1,4 @@
-"""Order-tracking analysis: matching, scoring, heuristic filters, and assembly."""
+"""Order-tracking analysis: matching, scoring, and finding assembly."""
 
 from __future__ import annotations
 
@@ -40,7 +40,11 @@ from vibesensor.use_cases.diagnostics.helpers import (
     _order_reference_spec_from_context,
     _sample_top_peaks,
 )
-from vibesensor.use_cases.diagnostics.math_utils import _mean
+from vibesensor.use_cases.diagnostics.order_heuristics import (
+    apply_localization_override,
+    detect_diffuse_excitation,
+    suppress_engine_aliases,
+)
 from vibesensor.use_cases.diagnostics.order_statistics import (
     compute_amplitude_and_error_stats,
     compute_matched_speed_phase_evidence,
@@ -250,166 +254,15 @@ def match_samples_for_hypothesis(
 
 
 # ═══════════════════════════════════════════════════════════════════════════
-# Scoring constants (local to heuristic filters in this module)
+# Multi-location split constants
 # ═══════════════════════════════════════════════════════════════════════════
 
-# ── Diffuse excitation detection constants ──────────────────────────────
-_DIFFUSE_AMPLITUDE_DOMINANCE_RATIO = 2.0
-_DIFFUSE_MATCH_RATE_RANGE_THRESHOLD = 0.15
-_DIFFUSE_MIN_MEAN_RATE = 0.15
-_DIFFUSE_PENALTY_BASE = 0.85
-_DIFFUSE_PENALTY_PER_SENSOR = 0.04
-_DIFFUSE_PENALTY_FLOOR = 0.65
-_HARMONIC_ALIAS_RATIO = 1.15
-_ENGINE_ALIAS_SUPPRESSION = 0.60
 # Maximum dominance ratio for splitting a finding into per-location findings.
 # A ratio of 2.0 means the secondary location must be at least 50% as strong.
 _MULTI_LOCATION_SPLIT_DOMINANCE = 2.0
 # Floor for dominance ratio when computing the secondary-location confidence
 # scale factor.  Prevents division by values at or below 1.0.
 _MIN_DOMINANCE_FOR_SCALE = 1.01
-
-
-def _normalized_source(finding: DomainFinding) -> str:
-    src: str = finding.source_normalized
-    return src
-
-
-def detect_diffuse_excitation(
-    connected_locations: set[str],
-    possible_by_location: dict[str, int],
-    matched_by_location: dict[str, int],
-    matched_points: list[OrderMatchObservation],
-    *,
-    min_match_points: int = ORDER_MIN_MATCH_POINTS,
-) -> tuple[bool, float]:
-    """Detect diffuse, non-localized excitation across multiple sensors."""
-    if len(connected_locations) < 2 or not possible_by_location:
-        return False, 1.0
-    loc_rates: list[float] = []
-    loc_mean_amps: dict[str, float] = {}
-    min_loc_points = max(3, min_match_points)
-    for location in connected_locations:
-        loc_possible = possible_by_location.get(location, 0)
-        loc_matched = matched_by_location.get(location, 0)
-        if loc_possible >= min_loc_points:
-            loc_rates.append(loc_matched / max(1, loc_possible))
-            loc_amps = [
-                point.amp
-                for point in matched_points
-                if (point.location or "").strip() == location and point.amp > 0
-            ]
-            if loc_amps:
-                loc_mean_amps[location] = _mean(loc_amps)
-    if len(loc_rates) < 2:
-        return False, 1.0
-    rate_range = max(loc_rates) - min(loc_rates)
-    mean_rate = _mean(loc_rates)
-    amp_uniform = True
-    if loc_mean_amps and len(loc_mean_amps) >= 2:
-        max_amp = max(loc_mean_amps.values())
-        min_amp = min(loc_mean_amps.values())
-        if min_amp > 0 and max_amp / min_amp > _DIFFUSE_AMPLITUDE_DOMINANCE_RATIO:
-            amp_uniform = False
-    if (
-        rate_range < _DIFFUSE_MATCH_RATE_RANGE_THRESHOLD
-        and mean_rate > _DIFFUSE_MIN_MEAN_RATE
-        and amp_uniform
-    ):
-        penalty = max(
-            _DIFFUSE_PENALTY_FLOOR,
-            _DIFFUSE_PENALTY_BASE - _DIFFUSE_PENALTY_PER_SENSOR * len(loc_rates),
-        )
-        return True, penalty
-    return False, 1.0
-
-
-def suppress_engine_aliases(
-    findings: list[tuple[float, DomainFinding]],
-    *,
-    min_confidence: float = ORDER_MIN_CONFIDENCE,
-) -> list[DomainFinding]:
-    """Suppress engine findings likely to be aliases of stronger wheel findings.
-
-    Engine findings whose ranking score is at least as high as the best wheel
-    finding are kept intact — a better ranking score indicates superior
-    frequency tracking, meaning the engine hypothesis is a better physical
-    explanation than the coincidentally close wheel harmonic.
-    """
-    best_wheel_conf = max(
-        (
-            finding.effective_confidence
-            for _, finding in findings
-            if _normalized_source(finding) == VibrationSource.WHEEL_TIRE
-        ),
-        default=0.0,
-    )
-    best_wheel_ranking = max(
-        (
-            score
-            for score, finding in findings
-            if _normalized_source(finding) == VibrationSource.WHEEL_TIRE
-        ),
-        default=0.0,
-    )
-    if best_wheel_conf > 0:
-        for index, (ranking_score, finding) in enumerate(findings):
-            if _normalized_source(finding) != VibrationSource.ENGINE:
-                continue
-            if ranking_score >= best_wheel_ranking:
-                continue
-            eng_conf = finding.effective_confidence
-            if eng_conf <= best_wheel_conf * _HARMONIC_ALIAS_RATIO:
-                suppressed = eng_conf * _ENGINE_ALIAS_SUPPRESSION
-                new_ranking_score = ranking_score * _ENGINE_ALIAS_SUPPRESSION
-                finding = replace(
-                    finding,
-                    confidence=suppressed,
-                    ranking_score=new_ranking_score,
-                )
-                findings[index] = (new_ranking_score, finding)
-    findings.sort(key=lambda item: item[0], reverse=True)
-    valid = [item[1] for item in findings if item[1].effective_confidence >= min_confidence]
-    return valid[:5]
-
-
-# ═══════════════════════════════════════════════════════════════════════════
-# Localization override
-# ═══════════════════════════════════════════════════════════════════════════
-
-
-def apply_localization_override(
-    *,
-    per_location_dominant: bool,
-    unique_match_locations: set[str],
-    connected_locations: set[str],
-    matched: int,
-    no_wheel_override: bool,
-    localization_confidence: float,
-    weak_spatial_separation: bool,
-    min_match_points: int = ORDER_MIN_MATCH_POINTS,
-) -> tuple[float, bool]:
-    """Adjust localization confidence when only one connected sensor matched."""
-    if (
-        per_location_dominant
-        and len(unique_match_locations) == 1
-        and len(connected_locations) >= 2
-        and not no_wheel_override
-    ):
-        localization_confidence = min(1.0, 0.50 + 0.15 * (len(connected_locations) - 1))
-        weak_spatial_separation = False
-    elif (
-        len(unique_match_locations) == 1
-        and len(connected_locations) >= 2
-        and matched >= min_match_points
-        and not no_wheel_override
-    ):
-        localization_confidence = max(
-            localization_confidence,
-            min(1.0, 0.40 + 0.10 * (len(connected_locations) - 1)),
-        )
-        weak_spatial_separation = False
-    return localization_confidence, weak_spatial_separation
 
 
 # ═══════════════════════════════════════════════════════════════════════════
